Organize GUI classes into package
