+++ src/main/java/cygnus/gui/MainWindow.java
@@ -1,5 +1,6 @@
-package cygnus;
+package cygnus.gui;
 
+import cygnus.Cygnus;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
